eCommerce_Task10-Adding new Movies into the list
Created a form that consists three input boxes- Movie Name, Movie Details, and Release Date, on the top of the fetch movie button to Add Movie. When the user clicks on the Add Movie button, show all that movie details as an object in the browser console by doing console.log(Movie). Also the input boxes should become clear again and display the newly added movie details in the same section where the fetch movie results are displayed.